Now compile on my iphone.

+++ source/expression.cpp
@@ -1,20 +1,20 @@
 /*---------------------------------------------------------------------------------
 
-    WordSphere Intepreter
-    Copyright (C) 2012 Quent42340 <quent42340@gmail.com>
+	WordSphere Intepreter
+	Copyright (C) 2012 Quent42340 <quent42340@gmail.com>
 
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
+	This program is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
 
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
 
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+	You should have received a copy of the GNU General Public License
+	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 ---------------------------------------------------------------------------------*/
 #include "header.h"
@@ -111,7 +111,7 @@ Value* OpExpression::evaluate() {
 				case '-': return new Value(val->value<int>() - val2->value<int>());
 				case '*': return new Value(val->value<int>() * val2->value<int>());
 				case '/': return new Value(val->value<int>() / val2->value<int>());
-				case '^': return new Value((int)pow(val->value<int>(), val2->value<int>()));
+				case '^': return new Value((int)pow((double)val->value<int>(), (double)val2->value<int>()));
 				case '%': return new Value(val->value<int>() % val2->value<int>());
 				case '<': return new Value(val->value<int>() < val2->value<int>());
 				case '>': return new Value(val->value<int>() > val2->value<int>());
